refactor(main): packages サービスを責務ごとに 5 ファイルへ分割
神モジュール化していた services/packages.ts(1143 行)を、承認済みの
Phase 5 設計(#2379 #A2-③)に沿って責務単位に分割する。挙動は同一で、
特性化テスト 28 件は無変更のまま分割後のモジュールを検証する。

- packageList: 一覧取得・状態計算・データ更新(getPackages 系 /
  refreshPackagesList / convertPackageIds / getIdDict)
- packageInstall: installPackageFlow / installPackageArchive /
  openPackageFolder
- packageUninstall: uninstallPackageFiles(+ removeScriptPackage)
- scriptInstall: installScriptFlow / installScriptArchive / getScriptsList
- packageShare: buildShareString / getEditorPackages / setEditorPackages

getPackagesWithStatus 内の fixIntegrity 分岐は
adoptManuallyInstalledPackages として切り出す(呼び出し順・挙動は維持)。
downloadRepository は設計表では scriptInstall 行だったが、利用元が
refreshPackagesList のみで表どおりだと packageList と scriptInstall の
循環 import になるため packageList に置く。
ARCHITECTURE.md はスライス①で反映漏れだった api/ 分割も含めて追随。
